INFRA_SCOPE/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy>=1.26.0
opencv-python==4.8.0.74
tensorflow==2.14.0
scikit-learn==1.3.2
Pillow==10.0.0
Werkzeug==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
matplotlib>=3.7.0
